source data overlay as component
#1119
