+++ waltz-ng/client/source-data-rating/index.js
@@ -4,7 +4,9 @@ export default (module) => {
         .config(require('./routes'))
         .service('SourceDataRatingStore', require('./services/source-data-rating-store'))
         .directive('waltzSourceDataInfo', require('./directives/source-data-info'))
-        .directive('waltzSourceDataOverlay', require('./directives/source-data-overlay'))
         .directive('waltzSourceDataSectionAddon', require('./directives/source-data-section-addon'));
+
+    module
+        .component('waltzSourceDataOverlay', require('./components/source-data-overlay'));
 
 };
